+++ Tests/Components/Renderers/TestLineRenderer.cs
@@ -2,5 +2,4 @@ namespace Termule.Tests.Components;
 
 public class TestLineRenderer
 {
-    
 }