+++ Tests/Core/TestSystemManager.cs
@@ -6,7 +6,15 @@ namespace Termule.Tests.Core;
 public class TestSystemManager
 {
     [Fact]
-    internal void Install_ShouldAddSystem()
+    public void Get_ShouldReturnNull_WhenSystemMissing()
+    {
+        var game = Game.Create();
+
+        Assert.Null(game.Systems.Get<FakeSystem>());
+    }
+
+    [Fact]
+    public void Install_ShouldAddSystem()
     {
         var game = Game.Create();
         FakeSystem system = new();
@@ -16,7 +24,7 @@ public class TestSystemManager
     }
 
     [Fact]
-    internal void Install_ShouldReplaceExistingSystem()
+    public void Install_ShouldReplaceExistingSystem()
     {
         var game = Game.Create();
         game.Systems.Install(new FakeSystem());
@@ -28,7 +36,7 @@ public class TestSystemManager
     }
 
     [Fact]
-    internal void Install_ShouldThrow_WhenGameStarted()
+    public void Install_ShouldThrow_WhenGameStarted()
     {
         var game = Game.Create();
         game.Prepare();
@@ -37,7 +45,7 @@ public class TestSystemManager
     }
 
     [Fact]
-    internal void InstalledSystem_ShouldFollowLifecycle()
+    public void InstalledSystem_ShouldFollowLifecycle()
     {
         var game = Game.Create();
         FakeSystem system = new();
@@ -54,7 +62,7 @@ public class TestSystemManager
     }
 
     [Fact]
-    internal void Uninstall_ShouldRemoveSystem()
+    public void Uninstall_ShouldRemoveSystem()
     {
         var game = Game.Create();
         game.Systems.Install(new FakeSystem());
@@ -65,19 +73,11 @@ public class TestSystemManager
     }
 
     [Fact]
-    internal void Uninstall_ShouldThrow_WhenGameStarted()
+    public void Uninstall_ShouldThrow_WhenGameStarted()
     {
         var game = Game.Create();
         game.Prepare();
 
         Assert.Throws<InvalidOperationException>(() => game.Systems.Uninstall<FakeSystem>());
-    }
-
-    [Fact]
-    internal void Get_ShouldReturnNull_WhenSystemMissing()
-    {
-        var game = Game.Create();
-
-        Assert.Null(game.Systems.Get<FakeSystem>());
     }
 }